requirements.txt
numpy
pandas
bs4
requests
apache-beam
google-cloud-pubsub
functions-framework==3.*
